cache test

+++ tests/cache/test.cpp
@@ -21,19 +21,16 @@ int main(int argc, char **argv)
 
 		coda_cache<int, int> cache (1024, 0);
 		cache.set(0, 23, false);
-
 		const int *v = cache.get(0);
-		printf("right after insert got %d\n", v ? *v : NULL);
+		printf("right after insert got %d\n", (int) (v ? *v : NULL));
 
 		sleep(1);
-
 		v = cache.get(0);
-		printf("one second later got %d\n", v ? *v : NULL);
+		printf("one second later got %d\n", (int) (v ? *v : NULL));
 
 		sleep(3);
-
 		v = cache.get(0);
-		printf("three more seconds later got %d\n", v ? *v : NULL);
+		printf("three more seconds later got %d\n", (int) (v ? *v : NULL));
 	}
 
 	/* TEST 2 */
@@ -42,27 +39,19 @@ int main(int argc, char **argv)
 
 		coda_cache<int, int> cache (1024, 1);
 		cache.set(0, 35, false);
-
 		const int *v = cache.get(0);
-		printf("right after insert got %d\n", v ? *v : NULL);
+		printf("right after insert got %d\n", (int) (v ? *v : NULL));
 
 		sleep(1);
-
 		v = cache.get(0);
-		printf("one second later got %d\n", v ? *v : NULL);
+		printf("one second later got %d\n", (int) (v ? *v : NULL));
 
 		sleep(3);
-
 		v = cache.get(0);
-		printf("three more seconds later got %d\n", v ? *v : NULL);
+		printf("three more seconds later got %d\n", (int) (v ? *v : NULL));
 	}
 
-
-
-
-
 #if 0
-
 	coda_cache<int, int> cache (1024 * 1024, 86400);
 
 	cache.set(0, 23, false);
@@ -85,7 +74,6 @@ int main(int argc, char **argv)
 	printf("sizeof(std::map<std::string,custom_type>::value_type)=%u\n", (unsigned) sizeof(std::map<std::string,custom_type>::value_type));
 	printf("sizeof(std::pair<std::string,custom_type>)=%u\n", (unsigned) sizeof(std::pair<std::string,custom_type>));
 #endif
-
 #endif
 
 	return 0;
